refactor(webapp_v2): tidy Configure components — fold SecretField states, drop dead code, fix placement
Make it obvious which Configure components are reusable, and remove what
the per-field description removal left unused.

- SecretField: fold the set / editing / new states (ReadOnlyStatus,
  ReplacingInput, NewInput) and FieldLabel into a single component driven
  by its state — variations over separate files. 8 files -> 4 (index,
  InlineAction, util, css).
- Dead code: removing per-field descriptions orphaned MarkdownText +
  parseMarkdownLinks and SourcedInput's `description` branch — delete them.
- Placement, per the convention (components/ = reusable, sections/ =
  single-consumer, page root = page chrome):
  * FormFooter -> page root (parity with ConfigureHeader).
  * ReviewSection, AIDataMaskingSection -> components/ (each used by two tabs).
  components/ is now exactly the reusable set: SecretField, ToggleSection,
  ReviewSection, AIDataMaskingSection.
- Update CLAUDE.md / COMPONENTS.md examples to match (the docs cited
  ReviewSection as a sections/ example, but it has two consumers).

No behavior change. Build + lint clean.
