+++ webapp_v2/src/pages/Roles/Configure/components/AIDataMaskingSection.jsx
@@ -6,7 +6,7 @@ import MultiSelect from '@/components/MultiSelect';
 import { useUserStore } from '@/stores/useUserStore';
 import { useConfigureRoleStore } from '../store';
 import { dlpOptionsFor } from '../utils/dlpInfoTypes';
-import ToggleSection from '../components/ToggleSection';
+import ToggleSection from './ToggleSection';
 
 // Shared AI Data Masking section used by both Terminal Access and
 // Native Access tabs. Gated by license and gateway capabilities, and
