+++ webapp_v2/src/pages/Roles/Configure/index.jsx
@@ -9,7 +9,7 @@ import { useBridgeStore } from '@/stores/useBridgeStore'
 import PageLoader from '@/components/PageLoader'
 import { useConfigureRoleStore } from './store'
 import ConfigureHeader from './ConfigureHeader'
-import FormFooter from './components/FormFooter'
+import FormFooter from './FormFooter'
 import CredentialsTab from './CredentialsTab'
 import DetailsTab from './DetailsTab'
 import TerminalAccessTab from './TerminalAccessTab'
